Add collector pkg & postgres SQL metric collector
Add the base structure of the metrics collector interface, that
would be form of a MetricsCollectorDriver which will provide
new Collector for each service instance and metrics source.

We implement a generic sqlMetricsCollectorDriver and
sqlMetricsCollector that from a given SQL driver, and connection
url, it can retrieve metrics from a predefined list of
Queries.

The list of Queries to retrieve the metrics is configurable. One
query can return multiple metrics. Each column is expected to
be named as the final key value of the metric exported. The
"unit" will be included in the resulting metrics.

The sqlMetricsCollectorDriver would use a `brokerinfo.BrokerInfo`
object to retrive the connection string to the DB, including
credentials.

We add the specific NewPostgresCollectorDriver() constructor that
will initialise the generic sqlMetricsCollectorDriver with postgres
specific queries.

We only add the number of connections for the time being.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>