Add pkg/scheduler to schedule metric collectors
The Scheduler will periodically query the list of instances
provided by the brokerinfo.BrokerInfo.

Whenever a new instance appears, it will use the given
MetricsCollectorDriver.NewCollector() to schedule a new worker
that will periodically collect the metrics of the give
MetricsCollector.

If any instance disappears in the list, the schedulled worker will
be stopped.

This package must be extended to support multiple
MetricsCollectorDriver (e.g. postgres, mysql, cloudwatch, etc.).

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>