Add pkg/brokerinfo AWS RDS implementation
Add an implementation for brokerInfo that queries the AWS RDS api
directly to retrieve the information, in a similar way than it is done
in the RDS Broker[1]. We copy across to list instances and
generate the password.

We rely on the packages from RDS broker awsrds[2], which couples
this code with the RDS broker itself. It also duplicates some logic
in the password generation.

We decided to go for this implementation as it was the quickest to
implement without refactoring the RDS broker.

Other approaches could be:
 - implement a secured endpoint in the RDS broker to retrieve the
   list of instances and the connection string.
 - refactor RDS broker to extract the shared logic in a module.

[1] https://github.com/alphagov/paas-rds-broker/blob/eee2df8257264e9afdbe9bc1b942174882e5d0d5/rdsbroker/broker.go#L611-L616
[2]  https://github.com/alphagov/paas-rds-broker/blob/eee2df8257264e9afdbe9bc1b942174882e5d0d5/rdsbroker/broker.go#L654-L670

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>