Add ci/blackbox for integration test
In this test we wil use the rds-broker service itself to provision
and deprovision an instance, and use a fake loggregator server
to the that the right messages are send properly.

The test will compile the rds-broker and the rds-metric-collector,
render the configuration accordingly and start one instance of
each for the whole suite.

Signed-off-by: Lee Porte <lee.porte@digital.cabinet-office.gov.uk>