Add pkg/emitter and loggregator emitter
Add the MetricsEmitter interface and implement and emitter to
send the metrics to logggegator. We use the upstream library
https://github.com/keymon/go-loggregator/

This code is based on the include example for the v2 protocol.

The library allows send Timers, Gauges and Deltas/Counters, but we
consider that for our use case Gauges is enough.

We copy the fake server tests from
https://github.com/cloudfoundry/go-loggregator/blob/8ebcfd3c7377510fe5a45ded5110d7749b562606/servers_test.go

We are not 100% coverage!!! we did not find a way to make the
first constructor fail in the test. :(

Signed-off-by: Lee Porte <lee.porte@digital.cabinet-office.gov.uk>