Implement CloudWatch collector
We'll be obtaining some metrics from the cloudwatch, as these may not be
accessible through the postgres instance. We've used counterfeiter to
generate our mocks and plan on using it there after.

We've used the following command to generate the mocks:

```sh
counterfeiter -o pkg/collector/mocks/cloudwatch.go vendor/github.com/aws/aws-sdk-go/service/cloudwatch/cloudwatchiface CloudWatchAPI
```

We use GetMetricStatistics to collect the metrics in the last minute
window for a period of 60 seconds. We might tune this in a follow up
commit. For more details see [1]

[1] https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_GetMetricStatistics.html