Add async collector creation
If a single collecor takes ages to create, for instance unreachable DB,
we'd like to carry on creating the collectors and eventually have the
one taking ages to catch up.

We also need to change the locking logic in the scheduler: we do not
want to stop if there is any worker being started or if the main
scheduler routing is checking for new instances. We use sync.WaitGroup
instead.

In the test we cannot rely on the job.IsRunning() anymore to
sync the test, as it runs too quickly.