Emit cloudwatch metrics perserving the timestamp
Cloudwatch metrics can have a big delay, from one to several minutes.
Because that we should emit our metrics to loggregator keeping the
timestamp reported by AWS.

We do this by adding a `Timestamp` field to the metrics. The
loggregator emitter would add this timestamp if it is != 0, or
use `time.Now()` if not.

The SQL collectors will not be modified and not populate the
timestamp, as the metric is emitted immediatelly when collected.

But the cloudwatch collector will retrieve the timestamp from the
cloudwatch datapoints and populate it accordingly.