Cloudwatch returns latest known metric in 10m
Cloudwatch might have some delay getting the metrics. We want to
query all the known the metrics in the last 10 minutes and
return the latest one.

We might consider change this behaviour later, for instance to
run the collector less often (e.g. every 10m) and send multiple
metrics since the last run.