Add tests for fake loggregator server
Add tests for the fake loggregator test.

We add one pending test for the case of receiving multiple envelopes.
Because some reason the fake server stops receiving envelopes after
a while, and we were not able to troubleshoot it.

This makes it complicated to extend the ci/blackbox tests using
this fake server.

Adding the test with a pending flag to work on the issue later.