Update concourse task to use golang:1.10
We added a piece of code that only works on > go1.8 (sort.Slice).

But the cf-acceptance-tests image uses go1.7. There is no need
to use the cf-acceptance-tests image, we can rely on the upstream
golang:1.10-alpine.

We need to install git and make as dependencies, but that is really
quick compared with all the integration test so not such a problem.