Refactor fake loggregator server
Make it more simple, as it will only send every received envelope
to a channel. We also allow receive multiple envelopes by running
a `for{}` in the BatchSend() gRPC method. Before it was receiving
only one message and stopping.

We rewrite all the tests (emmiter and blackbox) to use the new
implementation by checking the messages in the channel.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>