Ensure the service is stopped on signal
When receiving a TERM or INTERRUP signal, we want to be
sure that the process would first try to stop any running
worker and then exit after a timeout.

We do not want to block for ever so we add a timeout in the
signal handler in main. Receiving other signal will stop the
process immediatelly

Signed-off-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>