Refactor the scheduler.Stop logic
When we stop a worker in the scheduler, maybe it is running and
sending metrics. This was causing some flakiness in the scheduler
test. To solve it we add a loop in the StopWorker to wait in
case the job is running in that very moment.

We also refactor StartWorker and StopWorker to:
 - be private
 - spawn a go routine by themselves, so slow start/stop run
   in background.
 - keep track of running functions using wait groups

The scheduler stop will wait for the waitgroups to ensure that
the workers are stopped.

We also add a lock around the workers map for concurrency safety and
double check for nil pointers.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>