postgres: sql query for dbsize and test DB
We want to include the dbname tag with the current name, because
we might allow multiple DBs per service instance in the future.

Currently we only offer one DB per instance, which shall be returned
by the `brokerinfo.ConnectionString()`, so we only return only
`current_database()`.

We change the test to create a temporary test DB with a random DB
and delete it after the test. We had to kill all active connection
before DROP that database using SQL commands, because golang
sql.DB does keep a pool of connections and we did not find a way
to close them explicitely.