postgres: DB deadlocks counter metric
The deadlocks counter metric keeps a counter of how many deadlocks
have occurred per DB. A deadlock means that one transaction had to
be cancelled.

We add a test that simulates a deadlock.

We only report `current_database()` for the time being as our broker
only offers one DB per instance, but we might allow more later. We
include the tag `dbname`.

The example code to simulate the deadlock has been borrowed from:
https://www.depesz.com/2012/02/09/waiting-for-9-2-deadlock-counter/