postgres: seq_scan and idx_scan per table
We get the stats collector generated metric seq_scan and idx_scan
for each table. We add the tags for the table_name.

The dbname is forced to be `current_database()` but we might emit
for all the DBs if we offer more databases per instance.

We add a index to the films table and try to run queries that does
a scan and a index to check the counter.