Disable per-table and index metrics
To have per-table metrics is a nice to have, but not something
the user have requested. Also, we might have a explosion of metrics
when dealing with databases with lots of tables and indexes,
and this load is not easily predictable.

We have decided to disable this feature for the time being. The
tenants can query these metrics by themselves if they need to.

In this case, we:
 - remove per-table size
 - aggregate sequencial and index scan info