BrokerInfo uses a struct InstanceInfo
We will add additional drivers to our metrics collector,
specifically we want to add a Mysql one.

But because this we need to enrich the information retrieved
by the brokerinfo.BrokerInfo implementation with additional
information, instead of returning only the GUID string. With this
extra information we can later extend the `scheduler` to pick
one driver or other depending on the the extra information
of the instance.

In this commit we change the interface brokerinfo.BrokerInfo{} to
return and receive a struct InstanceInfo. This struct currently only
contains a GUID attribute and but we will extend it later.

We also do some minor refactor in the scheduler when stoping the
workers.