Collector Drivers for different instance types
We want to be able to instanciate some Collectors and not others
based on the instance type, so that we will be able to
add a specific Collectors for mysql but not postgres and viceversa.

In this commit we will add a new attribute to the
`brokerinfo.InstanceInfo`: `Type`, and extend the `CollectorDriver`
to report a list of supported types with `.SupportedTypes()`.

The scheduler will check if the list of supported types reported
by each driver contains the instance type of the instances
reported by brokerinfo.

We extend the existing collector drivers (sql and cloudwatch) and
we adapt the tests. We add the specific test to the scheduler.

In this commit everything would be hardcoded to support only
the `postgres` type, as it is the only implemented atm.