Add support for cloudwatch metrics for RDS mysql
Cloudwatch metrics for RDS postgres and mysql are the same[1]

We only need to report `mysql` as a supported type by the
CloudWatchCollector, and list both the `postgres` and `mysql`
instances in rdsbrokerinfo.

We are only reporting cloudwatch, there is no need to
implement the generation of the SQL connection string for mysql.

[1] https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/rds-metricscollected.html