Implement first MySQL SQL metric: connections
Implement the first basic metric for MySQL using SQL queries,
the number of connections.

We implement it as a query to the `performance_schema.global_status`
that is equivalent to SHOW STATUS but more versatile, as we can
do any SQL query (e.g. name columns). Note that this table is
only compatible with MySQL >= 5.7 [1]

The test is very similar to the one in postgresql and we reuse
several helpers, that we moved to helpers_test.go. The helper
injectDBName() required a minor update to support mysql urls.

We add the new driver to the scheduler instanced in main.go

We extend the Makefile to also start a mysql docker instance for
testing and update the README.md

In travis, we install mysql 5.7 as documented [2], setting the password
to empty string `""`. We had to set `sudo: true` to install mysql.

[1] https://dev.mysql.com/doc/refman/5.7/en/performance-schema-variable-table-migration.html
[2] https://docs.travis-ci.com/user/database-setup/#MySQL-57