Fix thread safe bug in scheduler.StopWorker()
The stopWorker() was not thread safe bacause it was accessing the
workers map from multiple threads without a lock.

If failed at least once in the blackbox test with:

         goroutine 3737 [running]:
         github.com/alphagov/paas-rds-metric-collector/pkg/scheduler.(*Scheduler).stopWorker.func1(0xc420129900, 0xb99798, 0x8, 0xc42055c1eb, 0x24)
         /go/src/github.com/alphagov/paas-rds-metric-collector/pkg/scheduler/scheduler.go:224 +0x498
         created by github.com/alphagov/paas-rds-metric-collector/pkg/scheduler.(*Scheduler).stopWorker
         /go/src/github.com/alphagov/paas-rds-metric-collector/pkg/scheduler/scheduler.go:205 +0x84

We refactor the scheduler to put all the workers map in a struct
with a lock for every operation, ensuring that any access is
thread safe.