Refactor metricQuery to be an interface
metricQuery is an interface with a method getMetrics()
which encapsulates how the metrics are extracted from
the query.

This way we can have multiple strategies to process the
query output.

Currently we have one implementation: columnMetricQuery
which returns one metric per column, with the key being
the column name:

```
mysql> SELECT
    ->     variable_value as connections
    -> FROM
    ->     performance_schema.global_status
    -> WHERE
    ->     variable_name = 'Threads_connected';
+-------------+
| connections |
+-------------+
| 1           |
+-------------+
1 row in set (0.01 sec)

```