MySQL connections use TLS but do not verify the host
The metrics collector lives on VMs without the AWS RDS CA root present.
This setup has worked fine for Postgres connections ("sslmode=require")
since the project's inception, hence should be acceptable for MySQL
connections. If it proves not to be acceptable, it should be solved for
both DB connection types simultaneously in a future story.