Create a mock locket server for tests
We'd like to test the interaction of the metrics collector program with
the locket server. We had some issues getting a real locket server to
start in the tests, so we decided to implement a mock server instead.

A locket server is a simple gRPC API which exposes four methods:

* Lock
* Release
* Fetch
* FetchAll

In our case the locket client only calls Lock repeatedly, so we can
simulate all of the required behaviour with a switch statement in that
method.

The idea is for tests to start up a mock locket server specifying the
behaviour they want with a command line flag. For example:

```
mockLocketServer := testhelpers.MockLocketServer{}
mockLocketServer.Build()
mockLocketServer.Run("../fixtures", "alwaysGrantLock")
```

...will compile a mock locket server and run it in such a way that it
always grants locks to clients. By running with different arguments it's
possible to make the server:

* alwaysGrantLock
* neverGrantLock
* grantLockAfterFiveAttempts
* grantLockOnceThenFail

Because it reaches into the fixtures directory to get its certs and
keys it also needs the path to that directory relative to the working
directory.