Acquire lock from locket before collecting metrics
We want multiple instances of rds-metric-collector to be able to run
simultaneously without reporting duplicate metrics. This will allow us
to run multiple instance for high availability or during blue-green
deployments.

This commit puts the collector schedule in an
[Ifrit](https://godoc.org/github.com/tedsuo/ifrit) process, which is
then added to an OrderedGroup containing a
[Locket](https://github.com/cloudfoundry/locket) process. The behaviour
is roughly:

* when the application starts the locket process attempts to acquire a
  lock every second
* when the locket process acquires the lock it reports itself ready to
  Ifrit by closing the `ready` channel
* Ifrit starts the collector process, which starts the scheduler
* the scheduler reports any errors which prevent it from starting to
  Ifrit, otherwise it reports itself ready and blocks until it receives
  a signal
* the locket process will exit with an error if it is unable to
  reacquire a lock within the 15 second TTL.
* Ifrit will send a signal to the collector process when the locket
  process errors
* the collector process will stop its workers and exit when it receives
  the signal from Ifrit
* the entire program will exit when the locket and collector processes
  have exited
* Monit will restart the rds-metrics-collector process, restarting the
  whole experience

We've added tests using a mock locket server which checks that the
process:

* should start cleanly if it can acquire a lock
* should hang if it cannot acquire a lock
* should hang until it acquires a lock, then start
* should crash if it loses the lock

Note that some of the tests are quite slow, because they have to wait
for things that poll once a second. These could be sped up significantly
if we made the poll interval configurable and set it to a lower value
when testing.