Add make tasks to start / stop mock locket server
Sadly killing the pid of `go run` only kills that process, but the
process we want to kill is a child. By storing the process group id of
`go run` we can use `kill -- -$PGID` which kills the entire group.

Storing the pid and logs in a tmp director, which I've added and ignored
with a .keep file.

Also update README / Makefile to document local running

Note that the `docker rm` command is not needed because we run `docker
run --rm` which cleans up after itself already.