Move ConnectionString logic SQL collectors
Move the logic to build the connection string to connect to
the specific DB to the SQL collectors for mysql and postgres.
We reduce the logic in the brokerinfo to simply return the connection
details provided by AWS RDS.

Each SQL collector would provide a specific implementation of
how to build the connection string based on the info provided
by the broker info.

This makes more sense, as the SQL logic is only in the collector
package, and it is more consistent when testing.