Continue on with the remaining metrics on error
At the current state, the for loop would stop the collection and return
error at any time when the error would occur.

This means, if the first metrics fails, due to timeout for instance,
none of the metric would be collected and dealt with.

Logging the error on the spot and not quitting, would assure these are
at least attempted and more errors would be printed to the log.