Inject the instance guid as a tag when emitting
We found out that when using the firehose exporter[1], the instance
ID was not being added as a label of the metric being exported.

Instead, all the metrics from all the instances have the same labels
and we can not make any use of these metrics in prometheus.

In this commit we add a new tag "guid" that should be translated
to new label in prometheus

[1] https://github.com/bosh-prometheus/firehose_exporter