Refactor scheduler to use simple timers
Do not use anymore a external library to schedule jobs in the
scheduler, instead use normal go routines with contexts to control
all the flow.

This way we have a unique go-routine for all the scheduler logic
and we can remove locks for the map and simplify the logic.

The collectorWorkers are now responsible of the initialisation of
the collector, and notify back if they stop using a channel. The
scheduler would keep track of the workers list.