Add retry logic to collector workers
Collection might fail due reasons. Currently we ignore the error and
wait for the next collection interval, but that means a big gap
between metrics, and maybe the error is not recoverable unless a new
collector is created.

We add logic to the worker to retry some times with exponential back off.

We set it as a default of 3 retries with a exponential backoff of
initially 1 second (so 1,2,4).