Add a go context to the collectors and collect metrics with a timeout
We have a problem with the SQL collectors when a Multi-AZ database is restarted
as the collector stays connected to the old IP address, but fails to time-out
for 15 minutes. We'll introduce a go context with a timeout to be used when
Collect() is called so we can recreate the old collector quicker.