Test max_tx_age and max_system_tx_age metrics
The new queries for max_tx_age and max_system_tx_age are centred around
checking that the user executing the query belongs (or doesn't, in the latter
case) to the application users group. This commit creates the group as a test
fixture and adds the postgres user to it, and tests transaction times with the
user both in and out of the group.