Copy the vendor dir to GOPATH for the integration tests
Also we need all files in the paas-rds-broker package

The integration tests build and run the rds-broker therefore we need the
repository for paas-rds-broker. Previously we simply ran a `go get` which
always checks out the latest master but during development we want the exact
same version running as we define in Gopkg.toml.