Install ginkgo CLI from vendor
This means that we get a consistent version of it instead of pulling
down the latest version from github each time.

As part of this, I've also changed the integration tests to use the
regular Go docker image instead of the alpine version, which makes the
setup a bit simpler.